Move tokenService to authService
