+++ src/main/kotlin/com/wafflestudio/account/api/domain/account/RefreshToken.kt
@@ -7,7 +7,7 @@ import java.time.LocalDateTime
 @Table("account_refresh_token")
 class RefreshToken(
     @Id
-    var id :Long? = null,
+    var id: Long? = null,
 
     val userId: Long,
