+++ src/main/kotlin/com/wafflestudio/account/api/domain/account/RefreshTokenRepository.kt
@@ -5,4 +5,5 @@ import org.springframework.stereotype.Repository
 
 @Repository
 interface RefreshTokenRepository: CoroutineCrudRepository<RefreshToken, Long> {
+    suspend fun findByToken(token: String): RefreshToken?
 }
