+++ src/main/kotlin/com/wafflestudio/account/api/error/ErrorType.kt
@@ -6,6 +6,10 @@ enum class ErrorType(
     val code: Int,
     val httpStatus: HttpStatus
 ) {
+
+    // 401
+    INVALID_TOKEN(1401001, HttpStatus.UNAUTHORIZED),
+
     // 403
     USER_INACTIVE(1403001, HttpStatus.FORBIDDEN),
