+++ src/main/kotlin/com/wafflestudio/account/api/interfaces/auth/AuthController.kt
@@ -1,8 +1,6 @@
 package com.wafflestudio.account.api.interfaces.auth
 
-import org.springframework.web.bind.annotation.PostMapping
-import org.springframework.web.bind.annotation.RequestBody
-import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.bind.annotation.*
 import javax.validation.Valid
 
 @RestController
@@ -14,5 +12,19 @@ class AuthController(
         @RequestBody @Valid signupRequest: SignupRequest,
     ): SignupResponse {
         return authService.signup(signupRequest)
+    }
+
+    @GetMapping("/v1/validate")
+    suspend fun tokenValidate(
+        @RequestBody @Valid validateRequest: ValidateRequest,
+    ): Unit {
+        return authService.validate(validateRequest)
+    }
+
+    @PutMapping("/v1/refresh")
+    suspend fun tokenRefresh(
+        @RequestBody @Valid refreshRequest: RefreshRequest,
+    ): RefreshResponse {
+        return authService.refresh(refreshRequest)
     }
 }
