--- src/main/kotlin/com/wafflestudio/account/api/interfaces/token/TokenRequest.kt
@@ -1,16 +1,0 @@
-package com.wafflestudio.account.api.interfaces.token
-
-import javax.validation.constraints.NotBlank
-
-data class ValidateRequest(
-        @field:NotBlank
-        val accessToken: String,
-)
-
-data class RefreshRequest(
-        @field:NotBlank
-        val accessToken: String,
-
-        @field:NotBlank
-        val refreshToken: String,
-)
